requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pypdf==4.0.1
groq==1.0.0